pyproject.toml
[project]
name = "legislation-scraper"
version = "0.1.0"
description = "Async web scraper for Brazilian federal and state legislation, with PDF/OCR extraction and structured data output."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=25.1.0",
    "aiohttp>=3.13.3",
    "aiohttp-socks>=0.11.0",
    "beautifulsoup4>=4.13.4",
    "certifi>=2026.1.4",
    "loguru>=0.7.3",
    "markitdown[all]>=0.1.5",
    "openai>=1.98.0",
    "playwright>=1.58.0",
    "playwright-stealth>=2.0.2",
    "pymupdf>=1.27.1",
    "python-dotenv>=1.1.1",
    "snowflake-connector-python>=4.3.0",
    "tenacity>=9.1.3",
    "tqdm>=4.67.3",
    "unidecode>=1.4.0",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "ipywidgets>=8.1.7",
    "jupyter>=1.1.1",
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.12.0",
]
